Improve the speed of AC-Tile Delta table creation
The StateQueue used to create Delta tables and failure tables in AC
MPM was checking for duplicates in the queue by walking the entire queue,
which is O(N^2). Instead, now allocate one byte for each state as a flag
to say it has already been visited. Only 1 bit is needed, but bytes are
simpler and slightly faster.

This could be applied to the other AC MPMs too. Could share the code for
StateQueue.