In AC-Tile, convert from using pids for indexing to pattern index
Use the MPM specific pattern index, which is simply an index starting at
zero and incrementing for each pattern added, rather than the externally
provided Pattern ID (pid), since that can be much larger than the number of
patterns. For example, one pattern with pid=8000 would result in a max_pid
of 8000, so the pid_pat_list would have 8000 entries.

The pid_pat_list[] becomes a list (array) of pattern indexes. The PID is
moved to the SCACTilePatternList as a single value. The PatternList is
also indexed by the Pattern Index.

max_pat_id is not longer needed and mpm_ctx->pattern_ctn is used instead.

The local bitarray is then also indexed by pattern index instead of PID, making
it much smaller too.