Keep SIDs lists for all MPMs sorted when they are created.
The code to add new sids to the array of matches SIDs requires that the
stored SIDs lists are sorted, since it does a merge sort of the new SIDs
matches into the found list.

TODO: The b2g, b3g and wumanber code resorts the stored SIDs list each
time a new SID is added, which is inefficient, but only affects startup
time. An insertion sort would be faster, or find one place in the prepare
function to do the sort.