Keep SIDs lists for all MPMs sorted when they are created.
The code to add new sids to the array of matches SIDs requires that the
stored SIDs lists are sorted, since it does a merge sort of the new SIDs
matches into the found list.

TODO: The b2g and b3g code resorts the stored SIDs list each time a new
SID is added, which is inefficient at startup.