Keep top-level bit mask for Cacheline blocks for pattern ID array bits
Create a new 64-bit mask for the PMQ pattern ID bit array such that
is the bit in the top-level mask is zero, all the bits in the corresponding
cache-line block are also zero (or treated as if they were zero). If the
bit is 1, then some bit in the cacheline block is a 1.

For 64 byte cache lines, each block is 8*64 = 512 bits. With 64 top-level
bits, that is 32K PIDS.

To clear the array, simply set top-level mask to zero.

The set and get accessors check the top-level bits first. The cachelines
are cleared on the first write.